pyproject.toml
[project]
name = "oagent"
version = "2024.0.0"
dependencies = [
  "asyncclick",
  "prompt_toolkit",
  "rich",
  "pyyaml",
  "ollama",
]
requires-python = ">=3.11"
authors = [
  {name = "Scott Pierce", email = "ddrscott@gmail.com"},
]
description = "Agentic Coder"
readme = "README.md"
classifiers = [
  "Programming Language :: Python"
]

[tool.setuptools.packages.find]
include = ["oagent"]

[project.scripts]
oagent = "oagent:cli.main"
